+++ pac4j-oauth/src/main/java/org/pac4j/oauth/profile/OAuth20Profile.java
@@ -14,27 +14,27 @@ public class OAuth20Profile extends CommonProfile {
     private static final long serialVersionUID = -2313972372691233648L;
 
     private transient static final String ACCESS_TOKEN = "access_token";
+    private transient static final String REFRESH_TOKEN = "refresh_token";
 
-    /**
-     * Set the access token
-     *
-     * @param accessToken the access token secret
-     */
     public void setAccessToken(final String accessToken) {
         addAttribute(ACCESS_TOKEN, accessToken);
     }
 
-    /**
-     * Return the access token.
-     *
-     * @return the access token
-     */
     public String getAccessToken() {
         return (String) getAttribute(ACCESS_TOKEN);
+    }
+
+    public void setRefreshToken(final String refreshToken) {
+        addAttribute(REFRESH_TOKEN, refreshToken);
+    }
+
+    public String getRefreshToken() {
+        return (String) getAttribute(REFRESH_TOKEN);
     }
 
     @Override
     public void removeLoginData() {
         removeAttribute(ACCESS_TOKEN);
+        removeAttribute(REFRESH_TOKEN);
     }
 }