+++ pac4j-oauth/src/main/java/org/pac4j/oauth/profile/creator/OAuthProfileCreator.java
@@ -87,7 +87,7 @@ abstract class OAuthProfileCreator implements ProfileCreator {
             throw new HttpCommunicationException("No data found for accessToken: " + accessToken);
         }
         final UserProfile profile = configuration.getProfileDefinition().extractUserProfile(body);
-        addAccessTokenToProfile(profile, accessToken);
+        addTokenToProfile(profile, accessToken);
         return Optional.of(profile);
     }
 
@@ -149,5 +149,5 @@ abstract class OAuthProfileCreator implements ProfileCreator {
      * @param profile     the user profile
      * @param accessToken the access token
      */
-    protected abstract void addAccessTokenToProfile(final UserProfile profile, final Token accessToken);
+    protected abstract void addTokenToProfile(final UserProfile profile, final Token accessToken);
 }