+++ pac4j-oauth/src/main/java/org/pac4j/oauth/profile/linkedin2/LinkedIn2ProfileCreator.java
@@ -40,7 +40,7 @@ public class LinkedIn2ProfileCreator extends OAuth20ProfileCreator {
             throw new HttpCommunicationException("Not data found for accessToken: " + accessToken);
         }
         final var profile = profileDefinition.extractUserProfile(body);
-        addAccessTokenToProfile(profile, accessToken);
+        addTokenToProfile(profile, accessToken);
 
         if (profile == null || !linkedin2Configuration.getScope().contains("r_emailaddress")) {
             return Optional.ofNullable(profile);